+++ src/main/java/nl/enjarai/showmeyourskin/mixin/ElytraFeatureRendererMixin.java
@@ -6,16 +6,20 @@ import net.minecraft.client.render.entity.feature.FeatureRenderer;
 import net.minecraft.client.render.entity.feature.FeatureRendererContext;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import nl.enjarai.showmeyourskin.config.ModConfig;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ElytraFeatureRenderer.class)
 public abstract class ElytraFeatureRendererMixin<T extends LivingEntity, M extends EntityModel<T>> extends FeatureRenderer<T, M> {
+    private final ThreadLocal<PlayerEntity> showmeyourskin$player = new ThreadLocal<>();
+
     public ElytraFeatureRendererMixin(FeatureRendererContext<T, M> context) {
         super(context);
     }
@@ -29,7 +33,26 @@ public abstract class ElytraFeatureRendererMixin<T extends LivingEntity, M exten
         if (livingEntity instanceof PlayerEntity player) {
             if (!ModConfig.INSTANCE.getApplicable(player.getUuid()).showElytra) {
                 ci.cancel();
+            } else {
+                showmeyourskin$player.set(player);
             }
         }
+    }
+
+    @ModifyArg(
+            method = "render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/entity/LivingEntity;FFFFFF)V",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/client/render/item/ItemRenderer;getArmorGlintConsumer(Lnet/minecraft/client/render/VertexConsumerProvider;Lnet/minecraft/client/render/RenderLayer;ZZ)Lnet/minecraft/client/render/VertexConsumer;"
+            ),
+            index = 3
+    )
+    private boolean showmeyourskin$hideElytraGlint(boolean original) {
+        var player = showmeyourskin$player.get();
+        if (player != null) {
+            return original && ModConfig.INSTANCE.getApplicable(player.getUuid()).getGlint(EquipmentSlot.CHEST);
+        }
+
+        return original;
     }
 }